requirements.txt
langgraph 
langgraph-api
langchain-openai 
langchain-groq 
python-dotenv
pychopg2-binary
pgvector
sentence-transformers
langchain-postgres